refactor: 抽出 PluginCard 精简插件管理界面，配置改弹窗+副标题并对齐设置项样式
