+++ src/components/settings/custom/PluginManager.vue
@@ -2,7 +2,7 @@
 import type { PluginInfo } from "@shared/types/plugin";
 import { usePluginsStore } from "@/stores/plugins";
 import { toast } from "@/composables/useToast";
-import { isExternalUrl, openExternal } from "@/utils/url";
+import { isExternalUrl } from "@/utils/url";
 
 defineOptions({ inheritAttrs: false });
 
@@ -23,32 +23,7 @@ onMounted(() => {
   if (!loaded.value) void pluginsStore.load();
 });
 
-type TagType = "default" | "success" | "info" | "warning" | "error";
-
-/** 当前状态 → 徽章颜色 + 文案 key */
-const statusTag = (info: PluginInfo): { type: TagType; key: string } => {
-  if (!info.enabled) return { type: "default", key: "disabled" };
-  switch (info.status.state) {
-    case "ready":
-      return { type: "success", key: "ready" };
-    case "loading":
-      return { type: "info", key: "loading" };
-    case "error":
-      return { type: "error", key: "error" };
-    case "disabled":
-      return { type: "default", key: "disabled" };
-    default:
-      return { type: "warning", key: "unloaded" };
-  }
-};
-
-/** ready 状态下展示声明的 sources 名 */
-const sourceNames = (info: PluginInfo): string[] => {
-  if (info.status.state !== "ready") return [];
-  return Object.keys(info.status.sources);
-};
-
-/** 控制类插件 ready 时的设置 schema */
+/** 控制类插件 ready 时声明的设置 schema（配置弹窗用） */
 const controlSettings = (info: PluginInfo) => {
   if (info.status.state !== "ready") return [];
   return info.status.settings ?? [];
@@ -152,7 +127,7 @@ const pendingName = computed(() => {
   );
 });
 
-/** 所有插件列表为空（两个分区均无插件） */
+/** 两个分区均无插件 */
 const isEmpty = computed(
   () => sourcePlugins.value.length === 0 && controlPlugins.value.length === 0,
 );
@@ -196,12 +171,8 @@ const isEmpty = computed(
       >
         <IconLucidePuzzle class="size-6" />
       </div>
-      <div class="text-sm text-on-surface-variant">
-        {{ t("settings.plugins.empty") }}
-      </div>
-      <div class="text-xs text-on-surface-variant/60">
-        {{ t("settings.plugins.emptyHint") }}
-      </div>
+      <div class="text-sm text-on-surface-variant">{{ t("settings.plugins.empty") }}</div>
+      <div class="text-xs text-on-surface-variant/60">{{ t("settings.plugins.emptyHint") }}</div>
     </div>
 
     <!-- 音源插件分区 -->
@@ -209,156 +180,16 @@ const isEmpty = computed(
       <div class="text-xs font-medium text-on-surface-variant/60 px-1">
         {{ t("settings.plugins.sectionSource") }}
       </div>
-      <ul class="flex flex-col gap-2.5 list-none p-0 m-0">
-        <li
+      <div class="flex flex-col gap-2.5">
+        <PluginCard
           v-for="info in sourcePlugins"
           :key="info.manifest.id"
-          class="rounded-xl bg-surface-panel border border-solid border-outline-variant/15 px-4 py-3.5"
-        >
-          <div class="flex items-start gap-3">
-            <!-- 主体信息 -->
-            <div class="flex-1 min-w-0">
-              <!-- 标题行：名字 + 版本 + 状态 + 外链 -->
-              <div class="flex items-center gap-2 flex-wrap">
-                <span class="text-sm font-medium text-on-surface truncate">
-                  {{ info.manifest.name }}
-                </span>
-                <button
-                  v-if="isExternalUrl(info.manifest.homepage)"
-                  type="button"
-                  class="p-0 border-none bg-transparent text-on-surface-variant/50 hover:text-primary cursor-pointer transition-colors leading-0"
-                  :title="info.manifest.homepage"
-                  @click="openExternal(info.manifest.homepage)"
-                >
-                  <IconLucideExternalLink class="size-3.5" />
-                </button>
-                <STag size="tiny" type="default" variant="soft">v{{ info.manifest.version }}</STag>
-                <STag size="tiny" :type="statusTag(info).type" variant="soft">
-                  {{ t(`settings.plugins.status.${statusTag(info).key}`) }}
-                </STag>
-                <STag v-if="info.updateInfo" size="tiny" type="warning" variant="soft">
-                  {{ t("settings.plugins.updateAvailable") }}
-                </STag>
-              </div>
-
-              <!-- 简介 -->
-              <div
-                v-if="info.manifest.description"
-                class="mt-1.5 flex items-start gap-1.5 text-xs text-on-surface-variant/70"
-              >
-                <IconLucideInfo class="size-3.5 shrink-0 mt-0.5 opacity-60" />
-                <span class="line-clamp-2">{{ info.manifest.description }}</span>
-              </div>
-
-              <!-- 作者 + 支持源 -->
-              <div
-                v-if="info.manifest.author || sourceNames(info).length"
-                class="mt-1.5 flex items-center gap-3 flex-wrap text-xs text-on-surface-variant/60"
-              >
-                <span v-if="info.manifest.author" class="flex items-center gap-1">
-                  <IconLucideUser class="size-3.5 opacity-60" />
-                  {{ info.manifest.author }}
-                </span>
-                <span v-if="sourceNames(info).length" class="flex items-center gap-1 flex-wrap">
-                  <STag
-                    v-for="src in sourceNames(info)"
-                    :key="src"
-                    size="tiny"
-                    variant="soft"
-                    type="primary"
-                    class="gap-1"
-                  >
-                    <IconLucideDatabase class="size-3" />
-                    {{ src }}
-                  </STag>
-                </span>
-              </div>
-
-              <!-- 错误信息 -->
-              <div
-                v-if="info.status.state === 'error'"
-                class="mt-2 rounded-md bg-red-500/10 px-2 py-1 text-xs text-red-500 break-all"
-              >
-                {{ info.status.error.message }}
-              </div>
-
-              <!-- 更新提示 -->
-              <div
-                v-if="info.updateInfo"
-                class="mt-2 rounded-md bg-amber-500/10 px-3 py-2 flex flex-col gap-1.5"
-              >
-                <div class="flex items-center justify-between gap-2">
-                  <div class="flex items-center gap-1.5 text-xs font-medium text-amber-600">
-                    <IconLucideArrowUpCircle class="size-3.5" />
-                    <span>
-                      {{ t("settings.plugins.newVersion") }}
-                      <template v-if="info.updateInfo.version">
-                        v{{ info.updateInfo.version }}
-                      </template>
-                    </span>
-                  </div>
-                  <SButton
-                    v-if="isExternalUrl(info.updateInfo.updateUrl)"
-                    variant="secondary"
-                    size="tiny"
-                    type="warning"
-                    @click="openExternal(info.updateInfo.updateUrl)"
-                  >
-                    <template #icon>
-                      <IconLucideExternalLink class="size-3" />
-                    </template>
-                    {{ t("settings.plugins.openUpdateUrl") }}
-                  </SButton>
-                </div>
-                <div
-                  v-if="info.updateInfo.log"
-                  class="text-xs text-on-surface-variant/80 whitespace-pre-wrap break-words"
-                >
-                  {{ info.updateInfo.log }}
-                </div>
-              </div>
-            </div>
-
-            <!-- 操作区 -->
-            <div class="shrink-0 flex items-center gap-2">
-              <SButton
-                v-if="controlSettings(info).length > 0"
-                variant="secondary"
-                size="small"
-                @click="openSettingsDialog(info.manifest.id)"
-              >
-                <template #icon>
-                  <IconLucideSettings2 class="size-4" />
-                </template>
-                {{ t("common.configure") }}
-              </SButton>
-              <SButton
-                :variant="info.enabled ? 'filled' : 'secondary'"
-                size="small"
-                :type="info.enabled ? 'primary' : 'default'"
-                @click="handleToggleEnabled(info.manifest.id, info.enabled)"
-              >
-                <template #icon>
-                  <IconLucideCircleCheck v-if="info.enabled" class="size-4" />
-                  <IconLucidePower v-else class="size-4" />
-                </template>
-                {{ info.enabled ? t("settings.plugins.enabled") : t("settings.plugins.enable") }}
-              </SButton>
-              <SButton
-                variant="secondary"
-                size="small"
-                type="error"
-                @click="openUninstallConfirm(info.manifest.id)"
-              >
-                <template #icon>
-                  <IconLucideTrash2 class="size-4" />
-                </template>
-                {{ t("settings.plugins.uninstall") }}
-              </SButton>
-            </div>
-          </div>
-        </li>
-      </ul>
+          :info="info"
+          @toggle="handleToggleEnabled"
+          @uninstall="openUninstallConfirm"
+          @configure="openSettingsDialog"
+        />
+      </div>
     </div>
 
     <!-- 控制插件分区 -->
@@ -366,143 +197,16 @@ const isEmpty = computed(
       <div class="text-xs font-medium text-on-surface-variant/60 px-1">
         {{ t("settings.plugins.sectionControl") }}
       </div>
-      <ul class="flex flex-col gap-2.5 list-none p-0 m-0">
-        <li
+      <div class="flex flex-col gap-2.5">
+        <PluginCard
           v-for="info in controlPlugins"
           :key="info.manifest.id"
-          class="rounded-xl bg-surface-panel border border-solid border-outline-variant/15 px-4 py-3.5"
-        >
-          <div class="flex items-start gap-3">
-            <!-- 主体信息 -->
-            <div class="flex-1 min-w-0">
-              <!-- 标题行：名字 + 版本 + 状态 + 外链 -->
-              <div class="flex items-center gap-2 flex-wrap">
-                <span class="text-sm font-medium text-on-surface truncate">
-                  {{ info.manifest.name }}
-                </span>
-                <button
-                  v-if="isExternalUrl(info.manifest.homepage)"
-                  type="button"
-                  class="p-0 border-none bg-transparent text-on-surface-variant/50 hover:text-primary cursor-pointer transition-colors leading-0"
-                  :title="info.manifest.homepage"
-                  @click="openExternal(info.manifest.homepage)"
-                >
-                  <IconLucideExternalLink class="size-3.5" />
-                </button>
-                <STag size="tiny" type="default" variant="soft">v{{ info.manifest.version }}</STag>
-                <STag size="tiny" :type="statusTag(info).type" variant="soft">
-                  {{ t(`settings.plugins.status.${statusTag(info).key}`) }}
-                </STag>
-                <STag v-if="info.updateInfo" size="tiny" type="warning" variant="soft">
-                  {{ t("settings.plugins.updateAvailable") }}
-                </STag>
-              </div>
-
-              <!-- 简介 -->
-              <div
-                v-if="info.manifest.description"
-                class="mt-1.5 flex items-start gap-1.5 text-xs text-on-surface-variant/70"
-              >
-                <IconLucideInfo class="size-3.5 shrink-0 mt-0.5 opacity-60" />
-                <span class="line-clamp-2">{{ info.manifest.description }}</span>
-              </div>
-
-              <!-- 作者 -->
-              <div
-                v-if="info.manifest.author"
-                class="mt-1.5 flex items-center gap-3 flex-wrap text-xs text-on-surface-variant/60"
-              >
-                <span class="flex items-center gap-1">
-                  <IconLucideUser class="size-3.5 opacity-60" />
-                  {{ info.manifest.author }}
-                </span>
-              </div>
-
-              <!-- 错误信息 -->
-              <div
-                v-if="info.status.state === 'error'"
-                class="mt-2 rounded-md bg-red-500/10 px-2 py-1 text-xs text-red-500 break-all"
-              >
-                {{ info.status.error.message }}
-              </div>
-
-              <!-- 更新提示 -->
-              <div
-                v-if="info.updateInfo"
-                class="mt-2 rounded-md bg-amber-500/10 px-3 py-2 flex flex-col gap-1.5"
-              >
-                <div class="flex items-center justify-between gap-2">
-                  <div class="flex items-center gap-1.5 text-xs font-medium text-amber-600">
-                    <IconLucideArrowUpCircle class="size-3.5" />
-                    <span>
-                      {{ t("settings.plugins.newVersion") }}
-                      <template v-if="info.updateInfo.version">
-                        v{{ info.updateInfo.version }}
-                      </template>
-                    </span>
-                  </div>
-                  <SButton
-                    v-if="isExternalUrl(info.updateInfo.updateUrl)"
-                    variant="secondary"
-                    size="tiny"
-                    type="warning"
-                    @click="openExternal(info.updateInfo.updateUrl)"
-                  >
-                    <template #icon>
-                      <IconLucideExternalLink class="size-3" />
-                    </template>
-                    {{ t("settings.plugins.openUpdateUrl") }}
-                  </SButton>
-                </div>
-                <div
-                  v-if="info.updateInfo.log"
-                  class="text-xs text-on-surface-variant/80 whitespace-pre-wrap break-words"
-                >
-                  {{ info.updateInfo.log }}
-                </div>
-              </div>
-            </div>
-
-            <!-- 操作区 -->
-            <div class="shrink-0 flex items-center gap-2">
-              <SButton
-                v-if="controlSettings(info).length > 0"
-                variant="secondary"
-                size="small"
-                @click="openSettingsDialog(info.manifest.id)"
-              >
-                <template #icon>
-                  <IconLucideSettings2 class="size-4" />
-                </template>
-                {{ t("common.configure") }}
-              </SButton>
-              <SButton
-                :variant="info.enabled ? 'filled' : 'secondary'"
-                size="small"
-                :type="info.enabled ? 'primary' : 'default'"
-                @click="handleToggleEnabled(info.manifest.id, info.enabled)"
-              >
-                <template #icon>
-                  <IconLucideCircleCheck v-if="info.enabled" class="size-4" />
-                  <IconLucidePower v-else class="size-4" />
-                </template>
-                {{ info.enabled ? t("settings.plugins.enabled") : t("settings.plugins.enable") }}
-              </SButton>
-              <SButton
-                variant="secondary"
-                size="small"
-                type="error"
-                @click="openUninstallConfirm(info.manifest.id)"
-              >
-                <template #icon>
-                  <IconLucideTrash2 class="size-4" />
-                </template>
-                {{ t("settings.plugins.uninstall") }}
-              </SButton>
-            </div>
-          </div>
-        </li>
-      </ul>
+          :info="info"
+          @toggle="handleToggleEnabled"
+          @uninstall="openUninstallConfirm"
+          @configure="openSettingsDialog"
+        />
+      </div>
     </div>
 
     <!-- URL 导入 -->
@@ -554,6 +258,7 @@ const isEmpty = computed(
     <SDialog
       v-model:open="settingsDialogOpen"
       :title="settingsDialogInfo?.manifest.name ?? ''"
+      :description="t('settings.plugins.configSubtitle')"
       width="520px"
     >
       <PluginSettingsForm
