+++ src/i18n/locales/zh-CN.json
@@ -1500,6 +1500,7 @@
       "enable": "启用",
       "enabled": "已启用",
       "uninstall": "卸载",
+      "configSubtitle": "插件配置",
       "uninstallConfirmTitle": "卸载插件",
       "uninstallConfirm": "确定要卸载插件「{name}」吗？此操作不可撤销。",
       "uninstallSuccess": "插件已卸载",
